refactor: renaming file and variable
