+++ seankwarren/react-flow-calc/src/utils/operationLabels.js
